allow for `display: contents` elements to directly return their contents
elements with `display: contents` have effectively no impact on layout. As far as the positioning of elements goes, the contents of a `display: contents` element belong to the parent element. Therefore, just directly return the children inline.

In order to accomodate this, parseNode now returns an array. For a normal node, this will just return an array with a single element. For an empty or not rendered node, it will return an empty array. For a `display: contents` node, it directly returns the parsed children.